Auto stash before merge of "sign" and "origin/sign"

+++ contracts/interfaces/IEmiVoting.sol
@@ -3,5 +3,5 @@
 pragma solidity ^0.6.2;
 
 interface IEmiVoting {
-  function getVotingResult(uint _hash) external view returns (address);
+    function getVotingResult(uint256 _hash) external view returns (address);
 }
